+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.11.0
+------------------------------
+*February 16, 2018*
+
+### Added
+- toggleSection / toggleAccordion / setToggleCallback methods
+
+### Changed
+- Structure of module to expose helper methods in helpers/external.js, and moved internal helpers to helpers/internal.js
+
 v0.10.0
 ------------------------------
 *January 29, 2018*